+++ src/app/Fake.DotNet.Cli/DotNet.fs
@@ -827,6 +827,86 @@ module DotNet =
         (fun opt -> { opt with DotNetCliPath = exe; Version = passVersion})
 
     /// dotnet restore command options
+    type MSBuildOptions =
+        {
+            /// Common tool options
+            Common: Options
+            MSBuildParams : MSBuild.CliArguments
+        }
+
+        /// Parameter default values.
+        static member Create() =
+          {  Common = Options.Create()
+             MSBuildParams = MSBuild.CliArguments.Create()
+          }
+
+        /// Gets the current environment
+        member x.Environment = x.Common.Environment
+        /// Sets the current environment variables.
+        member x.WithEnvironment map =
+            { x with Common = { x.Common with Environment = map } }
+
+        /// Sets a value indicating whether the output for the given process is redirected.
+        member x.WithRedirectOutput shouldRedirect =
+            { x with Common = x.Common.WithRedirectOutput shouldRedirect }
+
+        /// Changes the "Common" properties according to the given function
+        member inline x.WithCommon f =
+            { x with Common = f x.Common }
+
+    let internal addBinaryLogger disableFakeBinLog args (common:Options) =
+        // used for detection
+        let callMsBuildExe args =
+            let result =
+                exec (fun _ -> 
+                    { common with
+                       RedirectOutput = true }) "msbuild" args
+            if not result.OK then
+                failwithf "msbuild failed with exitcode '%d'" result.ExitCode
+            String.Join("\n", result.Messages)
+        MSBuild.addBinaryLogger (common.DotNetCliPath + " msbuild") callMsBuildExe args disableFakeBinLog
+   
+    let internal execWithBinLog project common command args msBuildArgs =
+        let argString = MSBuild.fromCliArguments msBuildArgs
+        let binLogPath, args = addBinaryLogger msBuildArgs.DisableInternalBinLog (args + " " + argString) common
+        let result = exec (fun _ -> common) command args
+        MSBuild.handleAfterRun (sprintf "dotnet %s" command) binLogPath result.ExitCode project
+
+    /// Runs a MSBuild project
+    /// ## Parameters
+    ///  - `setParams` - A function that overwrites the default MSBuildOptions
+    ///  - `project` - A string with the path to the project file to build.
+    ///
+    /// ## Sample
+    ///
+    ///     open Fake.DotNet
+    ///     let setMsBuildParams (defaults:MSBuild.CliArguments) =
+    ///             { defaults with
+    ///                 Verbosity = Some(Quiet)
+    ///                 Targets = ["Build"]
+    ///                 Properties =
+    ///                     [
+    ///                         "Optimize", "True"
+    ///                         "DebugSymbols", "True"
+    ///                         "Configuration", "Release"
+    ///                     ]
+    ///              }
+    ///     let setParams (defaults:DotNet.MSBuildOptions) =
+    ///             { defaults with
+    ///                 MSBuildParams = setMsBuildParams defaults.MSBuildParams
+    ///              }
+    ///     
+    ///     DotNet.msbuild setParams "./MySolution.sln"
+    let msbuild setParams project =
+        use __ = Trace.traceTask "DotNet:msbuild" project
+        
+        let param = MSBuildOptions.Create() |> setParams
+        let args = [project]
+        let args = Args.toWindowsCommandLine args
+        execWithBinLog project param.Common "msbuild" args param.MSBuildParams
+        __.MarkSuccess()
+
+    /// dotnet restore command options
     type RestoreOptions =
         {
             /// Common tool options
@@ -845,6 +925,8 @@ module DotNet =
             IgnoreFailedSources: bool
             /// Disables restoring multiple projects in parallel (--disable-parallel)
             DisableParallel: bool
+            /// Other msbuild specific parameters
+            MSBuildParams : MSBuild.CliArguments
         }
 
         /// Parameter default values.
@@ -857,6 +939,7 @@ module DotNet =
             NoCache = false
             IgnoreFailedSources = false
             DisableParallel = false
+            MSBuildParams = MSBuild.CliArguments.Create()
         }
         [<Obsolete("Use Options.Create instead")>]
         static member Default = Options.Create()
@@ -897,9 +980,8 @@ module DotNet =
     let restore setParams project =
         use __ = Trace.traceTask "DotNet:restore" project
         let param = RestoreOptions.Create() |> setParams
-        let args = project :: buildRestoreArgs param
-        let result = exec (fun _ -> param.Common) "restore" (Args.toWindowsCommandLine args)
-        if not result.OK then failwithf "dotnet restore failed with code %i" result.ExitCode
+        let args = Args.toWindowsCommandLine(project :: buildRestoreArgs param)
+        execWithBinLog project param.Common "restore" args param.MSBuildParams
         __.MarkSuccess()
 
     /// build configuration
@@ -907,7 +989,6 @@ module DotNet =
         | Debug
         | Release
         | Custom of string
-    with
         /// Convert the build configuration to a string that can be passed to the .NET CLI
         override this.ToString() =
             match this with
@@ -950,6 +1031,8 @@ module DotNet =
             OutputPath: string option
             /// No build flag (--no-build)
             NoBuild: bool
+            /// Other msbuild specific parameters
+            MSBuildParams : MSBuild.CliArguments
         }
 
         /// Parameter default values.
@@ -960,6 +1043,7 @@ module DotNet =
             BuildBasePath = None
             OutputPath = None
             NoBuild = false
+            MSBuildParams = MSBuild.CliArguments.Create()
         }
         [<Obsolete("Use PackOptions.Create instead")>]
         static member Default = PackOptions.Create()
@@ -997,9 +1081,8 @@ module DotNet =
     let pack setParams project =
         use __ = Trace.traceTask "DotNet:pack" project
         let param = PackOptions.Create() |> setParams
-        let args = project :: buildPackArgs param
-        let result = exec (fun _ -> param.Common) "pack" (Args.toWindowsCommandLine args)
-        if not result.OK then failwithf "dotnet pack failed with code %i" result.ExitCode
+        let args = Args.toWindowsCommandLine(project :: buildPackArgs param)
+        execWithBinLog project param.Common "pack" args param.MSBuildParams
         __.MarkSuccess()
 
     /// dotnet publish command options
@@ -1021,6 +1104,8 @@ module DotNet =
             VersionSuffix: string option
             /// No build flag (--no-build)
             NoBuild: bool
+            /// Other msbuild specific parameters
+            MSBuildParams : MSBuild.CliArguments
         }
 
         /// Parameter default values.
@@ -1033,6 +1118,7 @@ module DotNet =
             OutputPath = None
             VersionSuffix = None
             NoBuild = false
+            MSBuildParams = MSBuild.CliArguments.Create()
         }
         [<Obsolete("Use PublishOptions.Create instead")>]
         static member Default = PublishOptions.Create()
@@ -1072,9 +1158,8 @@ module DotNet =
     let publish setParams project =
         use __ = Trace.traceTask "DotNet:publish" project
         let param = PublishOptions.Create() |> setParams
-        let args = project :: buildPublishArgs param
-        let result = exec (fun _ -> param.Common) "publish" (Args.toWindowsCommandLine args)
-        if not result.OK then failwithf "dotnet publish failed with code %i" result.ExitCode
+        let args = Args.toWindowsCommandLine(project :: buildPublishArgs param)
+        execWithBinLog project param.Common "publish" args param.MSBuildParams
         __.MarkSuccess()
 
     /// dotnet build command options
@@ -1094,6 +1179,8 @@ module DotNet =
             OutputPath: string option
             /// Native flag (--native)
             Native: bool
+            /// Other msbuild specific parameters
+            MSBuildParams : MSBuild.CliArguments
         }
 
         /// Parameter default values.
@@ -1105,6 +1192,7 @@ module DotNet =
             BuildBasePath = None
             OutputPath = None
             Native = false
+            MSBuildParams = MSBuild.CliArguments.Create()
         }
         [<Obsolete("Use BuildOptions.Create instead")>]
         static member Default = BuildOptions.Create()
@@ -1144,9 +1232,8 @@ module DotNet =
     let build setParams project =
         use __ = Trace.traceTask "DotNet:build" project
         let param = BuildOptions.Create() |> setParams
-        let args = project :: buildBuildArgs param
-        let result = exec (fun _ -> param.Common) "build" (Args.toWindowsCommandLine args)
-        if not result.OK then failwithf "dotnet build failed with code %i" result.ExitCode
+        let args = Args.toWindowsCommandLine(project :: buildBuildArgs param)
+        execWithBinLog project param.Common "build" args param.MSBuildParams
         __.MarkSuccess()
 
     /// dotnet build command options
@@ -1187,6 +1274,8 @@ module DotNet =
             NoRestore: bool
             /// Arguments to pass runsettings configurations through commandline. Arguments may be specified as name-value pair of the form [name]=[value] after "-- ". Note the space after --.
             RunSettingsArguments : string option
+            /// Other msbuild specific parameters
+            MSBuildParams : MSBuild.CliArguments
         }
 
         /// Parameter default values.
@@ -1206,6 +1295,7 @@ module DotNet =
             Collect = None
             NoRestore = false
             RunSettingsArguments = None
+            MSBuildParams = MSBuild.CliArguments.Create()
         }
         [<Obsolete("Use TestOptions.Create instead")>]
         static member Default = TestOptions.Create()
@@ -1251,8 +1341,7 @@ module DotNet =
     let test setParams project =
         use __ = Trace.traceTask "DotNet:test" project
         let param = TestOptions.Create() |> setParams
-        let args = project :: buildTestArgs param
-        let result = exec (fun _ -> param.Common) "test" (Args.toWindowsCommandLine args)
-        if not result.OK then failwithf "dotnet test failed with code %i" result.ExitCode
+        let args = Args.toWindowsCommandLine(project :: buildTestArgs param)
+        execWithBinLog project param.Common "test" args param.MSBuildParams
         __.MarkSuccess()