+++ src/test/Fake.Core.UnitTests/Fake.DotNet.MSBuild.fs
@@ -8,10 +8,11 @@ open Expecto
 let tests =
   testList "Fake.DotNet.MSBuild.Tests" [
     testCase "Test that we can create simple msbuild cmdline" <| fun _ ->
-      let cmdLine =
-        { MSBuildParams.Create() with
-            Properties = ["OutputPath", "C:\\Test\\"] }
-        |> MSBuild.serializeMSBuildParams
+      let _, cmdLine =
+        MSBuild.buildArgs (fun defaults ->
+          { defaults with
+              ConsoleLogParameters = []
+              Properties = ["OutputPath", "C:\\Test\\"] })
       let expected =
         if Environment.isUnix then "\"/p:RestorePackages=False\" \"/p:OutputPath=C:\\Test\\\\\""    
         else "\"/m\" \"/nodeReuse:False\" \"/p:RestorePackages=False\" \"/p:OutputPath=C:\\Test\\\\\""